move to proper areas
