+++ src/Tk.Extensions/Io/IoExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using System.Diagnostics.CodeAnalysis;
+using Tk.Extensions.Guards;
 
 namespace Tk.Extensions.Io
 {
